feat: use event layout in timeline
